MAGECLOUD-4101: Introduce scenarios processor and parser mechanism
