+++ composer.json
@@ -22,7 +22,7 @@
     "illuminate/config": "^5.6",
     "illuminate/container": "^5.6",
     "illuminate/contracts": "^5.6",
-    "magento/magento-cloud-components": "^1.0",
+    "magento/magento-cloud-components": "^1.0.1",
     "monolog/monolog": "^1.16",
     "psr/container": "^1.0",
     "psr/log": "^1.0",
